Added SeedData and connected Apollo
Since we have a backend and frontend, we need something to glue them together - that's Apollo (the middlewear). Next steps are to add hooks to fetch data and start displaying things.

A bit confused about the code written on `_app.js`, but hopefully it will start to make more sense as we go down the line.